Fix test case failure due to package move.
