+++ azure-maven-plugin-common/src/test/java/com/microsoft/azure/maven/common/TestHelper.java
@@ -7,7 +7,13 @@
 package com.microsoft.azure.maven.common;
 
 import com.google.common.base.Preconditions;
+import org.apache.maven.model.Model;
+import org.apache.maven.model.io.DefaultModelReader;
+import org.apache.maven.model.io.ModelReader;
 
+import java.io.File;
+import java.io.IOException;
+import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -15,5 +21,10 @@ public class TestHelper {
     public static String joinIntegers(List<Integer> integers) {
         Preconditions.checkNotNull(integers, "Parameter 'integers' should not be null or empty.");
         return integers.stream().map(t->t.toString()).collect(Collectors.joining(","));
+    }
+
+    public static Model readMavenModel(File pomFile) throws IOException {
+        final ModelReader reader = new DefaultModelReader();
+        return reader.read(pomFile, Collections.emptyMap());
     }
 }
