+++ azure-maven-plugin-common/src/test/java/com/microsoft/azure/maven/common/utils/MavenUtilsTest.java
@@ -6,8 +6,7 @@
 
 package com.microsoft.azure.maven.common.utils;
 
-import com.microsoft.azure.maven.common.utils.MavenUtils;
-import com.microsoft.azure.maven.spring.TestHelper;
+import com.microsoft.azure.maven.common.TestHelper;
 import org.apache.maven.model.Model;
 import org.apache.maven.project.MavenProject;
 import org.codehaus.plexus.util.xml.Xpp3Dom;
